+++ Editor/Discovery.cs
@@ -13,11 +13,13 @@ namespace Microsoft.Unity.VisualStudio.Editor
 	{
 		public static IEnumerable<IVisualStudioInstallation> GetVisualStudioInstallations()
 		{
+#if UNITY_EDITOR_WIN
 			foreach (var installation in VisualStudioForWindowsInstallation.GetVisualStudioInstallations())
 				yield return installation;
-
+#elif UNITY_EDITOR_OSX
 			foreach (var installation in VisualStudioForMacInstallation.GetVisualStudioInstallations())
 				yield return installation;
+#endif
 
 			foreach (var installation in VisualStudioCodeInstallation.GetVisualStudioInstallations())
 				yield return installation;
@@ -27,12 +29,13 @@ namespace Microsoft.Unity.VisualStudio.Editor
 		{
 			try
 			{
+#if UNITY_EDITOR_WIN
 				if (VisualStudioForWindowsInstallation.TryDiscoverInstallation(editorPath, out installation))
 					return true;
-
+#elif UNITY_EDITOR_OSX
 				if (VisualStudioForMacInstallation.TryDiscoverInstallation(editorPath, out installation))
 					return true;
-
+#endif
 				if (VisualStudioCodeInstallation.TryDiscoverInstallation(editorPath, out installation))
 					return true;
 			}
@@ -46,8 +49,11 @@ namespace Microsoft.Unity.VisualStudio.Editor
 
 		public static void Initialize()
 		{
+#if UNITY_EDITOR_WIN
 			VisualStudioForWindowsInstallation.Initialize();
+#elif UNITY_EDITOR_OSX
 			VisualStudioForMacInstallation.Initialize();
+#endif
 			VisualStudioCodeInstallation.Initialize();
 		}
 	}