+++ Editor/ProjectGeneration/LegacyStyleProjectGeneration.cs
@@ -12,6 +12,8 @@ namespace Microsoft.Unity.VisualStudio.Editor
 
 	internal class LegacyStyleProjectGeneration : ProjectGeneration
 	{
+		internal override string StyleName => "Legacy";
+
 		public LegacyStyleProjectGeneration(string tempDirectory, IAssemblyNameProvider assemblyNameProvider, IFileIO fileIoProvider, IGUIDGenerator guidGenerator) : base(tempDirectory, assemblyNameProvider, fileIoProvider, guidGenerator)
 		{
 		}
@@ -82,7 +84,7 @@ namespace Microsoft.Unity.VisualStudio.Editor
 		internal override void GetProjectFooter(StringBuilder footerBuilder)
 		{
 			footerBuilder.Append(string.Join(k_WindowsNewline,
-				@"  <Import Project=""$(MSBuildToolsPath)\Microsoft.CSharp.targets"" />",
+				$"  <Import Project=\"{@"$(MSBuildToolsPath)\Microsoft.CSharp.targets".NormalizePathSeparators()}\" />",
 				@"  <Target Name=""GenerateTargetFrameworkMonikerAttribute"" />",
 				@"  <!-- To modify your build process, add your task inside one of the targets below and uncomment it.",
 				@"       Other similar extension points exist, see Microsoft.Common.targets.",